extract sync lookup utils
